third_party/blink/web_tests/http/tests/media/autoplay/document-user-activation-navigation-nogesture-expected.txt

CONSOLE WARNING: The AudioContext was not allowed to start. It must be resumed (or created) after a user gesture on the page. https://goo.gl/7K7WLu
FAIL NotAllowedError: play() failed because the user didn't interact with the document first. https://goo.gl/xX8pDD should be undefined. Was NotAllowedError: play() failed because the user didn't interact with the document first. https://goo.gl/xX8pDD
FAIL results.webAudio should be running. Was suspended.
PASS successfullyParsed is true

TEST COMPLETE


============== Back Forward List ==============
curr->  http://127.0.0.1:8000/media/autoplay/resources/test-autoplay.html
===============================================
